Create counting guidelines
This adds in guidelines on how to send messages in the counting channel. 

Currently, the counting channel is governed both by unspoken rules, and by Adam's bot, which is created by those rules, but also added some rules by its existence. It may be in the interest of active people to have a defined ruleset that:

* Can be shown to people to get them into counting
* Is a way to let already-active people know of unspoken rules they may not know of